alteração no formato de remover as notificações
