+++ scripts/V51__INCLUSAO_CAMPO_EXCLUIDO_NOTIFICACOES.sql
@@ -1,0 +1,3 @@
+ALTER TABLE if exists public.notificacao ADD if not exists excluido boolean not null default false;
+ALTER TABLE if exists public.notificacao_aluno ADD if not exists excluido boolean not null default false;
+ALTER TABLE if exists public.notificacao_turma ADD if not exists excluido boolean not null default false;
