+++ src/SME.AE.Api/Program.cs
@@ -20,6 +20,6 @@ namespace SME.AE.Api
             })
             .UseStartup<Startup>()
             .UseSentry()
-            .UseUrls("http://0.0.0.0:5000;");
+            .UseUrls("http://0.0.0.0:5030;");
     }
 }
