+++ src/SME.AE.Aplicacao/Comandos/Notificacao/RemoverNotificacao/RemoverNotificacaoCommand.cs
@@ -1,0 +1,14 @@
+﻿using MediatR;
+
+namespace SME.AE.Aplicacao.Comandos.Notificacao.Remover
+{
+    public class RemoverNotificacaoCommand : IRequest<string[]>
+    {
+        public long[] Ids { get; set; }
+
+        public RemoverNotificacaoCommand(long[] ids)
+        {
+            Ids = ids;
+        }
+    }
+}
