+++ src/SME.AE.Aplicacao/Comandos/Notificacao/RemoverNotificacaoPorId/RemoverNotificacaoPorIdCommand.cs
@@ -1,0 +1,14 @@
+﻿using MediatR;
+
+namespace SME.AE.Aplicacao.Comandos.Notificacao.Remover
+{
+    public class RemoverNotificacaoPorIdCommand : IRequest<bool>
+    {
+        public int Id { get; set; }
+
+        public RemoverNotificacaoPorIdCommand(int id)
+        {
+            Id = id;
+        }
+    }
+}
