+++ src/SME.AE.Aplicacao/Comandos/Notificacao/RemoverNotificacaoPorId/RemoverNotificacaoPorIdCommandHandler.cs
@@ -1,29 +1,15 @@
-﻿using System.Collections.Generic;
-using System.Text;
+﻿using MediatR;
+using SME.AE.Aplicacao.Comum.Interfaces.Repositorios;
 using System.Threading;
 using System.Threading.Tasks;
-using MediatR;
-using Microsoft.AspNetCore.Mvc;
-using SME.AE.Aplicacao.Comum.Interfaces.Repositorios;
-using SME.AE.Aplicacao.Comum.Modelos;
 
 namespace SME.AE.Aplicacao.Comandos.Notificacao.Remover
 {
-    public class RemoverNotificacaoPorIdCommand : IRequest<bool>
-    {
-        public int Id { get; set; }
-
-        public RemoverNotificacaoPorIdCommand(int id)
-        {
-            Id = id;
-        }
-    }
-
     public class RemoverNotificacaoPorIdCommandHandler : IRequestHandler<RemoverNotificacaoPorIdCommand, bool>
     {
-        private readonly INotificacaoRepository _repository;
+        private readonly INotificacaoRepositorio _repository;
 
-        public RemoverNotificacaoPorIdCommandHandler(INotificacaoRepository repository)
+        public RemoverNotificacaoPorIdCommandHandler(INotificacaoRepositorio repository)
         {
             _repository = repository;
         }
