+++ src/SME.AE.Aplicacao/Comandos/Notificacao/RemoverNotificacaoUsuario/RemoverNotificacaoUsuarioCommand.cs
@@ -1,0 +1,14 @@
+﻿using MediatR;
+
+namespace SME.AE.Aplicacao.Comandos.Notificacao.Remover
+{
+    public class RemoverNotificacaoUsuarioCommand : IRequest<bool>
+    {
+        public long[] Ids { get; set; }
+
+        public RemoverNotificacaoUsuarioCommand(long[] ids)
+        {
+            Ids = ids;
+        }
+    }
+}
