+++ src/SME.AE.Aplicacao/Comandos/Notificacao/RemoverNotificacaoUsuario/RemoverNotificacaoUsuarioCommandHandler.cs
@@ -1,23 +1,10 @@
 ﻿using MediatR;
 using SME.AE.Aplicacao.Comum.Interfaces.Repositorios;
-using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
 namespace SME.AE.Aplicacao.Comandos.Notificacao.Remover
 {
-    public class RemoverNotificacaoUsuarioCommand : IRequest<bool>
-    {
-        public long[] Ids { get; set; }
-
-        public RemoverNotificacaoUsuarioCommand(long[] ids)
-        {
-            Ids = ids;
-        }
-    }
-
     public class RemoverNotificacaoUsuarioCommandHandler : IRequestHandler<RemoverNotificacaoUsuarioCommand, bool>
     {
         private readonly IUsuarioNotificacaoRepositorio _repository;
@@ -34,7 +21,7 @@ namespace SME.AE.Aplicacao.Comandos.Notificacao.Remover
             {
                 bool retorno = await _repository.Remover(request.Ids[i]);
 
-               if (!retorno)
+                if (!retorno)
                     return false;
             }
