+++ src/SME.AE.Aplicacao/Consultas/Notificacao/MensagenUsuarioLogadoAlunoId/MensagenUsuarioLogadoAlunoIdQueryHandler.cs
@@ -13,10 +13,10 @@ namespace SME.AE.Aplicacao.Consultas.Notificacao.ListarNotificacaoAluno
 {
     public class MensagenUsuarioLogadoAlunoIdQueryHandler : IRequestHandler<MensagenUsuarioLogadoAlunoIdQuery, NotificacaoResposta>
     {
-        private readonly INotificacaoRepository notificacaoRepository;
+        private readonly INotificacaoRepositorio notificacaoRepository;
         private readonly IGrupoComunicadoRepository grupoComunicadoRepository;
 
-        public MensagenUsuarioLogadoAlunoIdQueryHandler(INotificacaoRepository notificacaoRepository, IGrupoComunicadoRepository grupoComunicadoRepository)
+        public MensagenUsuarioLogadoAlunoIdQueryHandler(INotificacaoRepositorio notificacaoRepository, IGrupoComunicadoRepository grupoComunicadoRepository)
         {
             this.notificacaoRepository = notificacaoRepository ?? throw new ArgumentNullException(nameof(notificacaoRepository));
             this.grupoComunicadoRepository = grupoComunicadoRepository ?? throw new ArgumentNullException(nameof(grupoComunicadoRepository));
